feat: move block domain under user package
