+++ src/main/kotlin/org/appjam/smashing/domain/user/dto/command/UserBlockCommand.kt
@@ -1,4 +1,4 @@
-package org.appjam.smashing.domain.block.dto.command
+package org.appjam.smashing.domain.user.dto.command
 
 data class UserBlockCommand(
     val blockedUserProfileId: String,
