+++ src/main/kotlin/org/appjam/smashing/domain/user/repository/BlockRepositoryCustom.kt
@@ -1,4 +1,4 @@
-package org.appjam.smashing.domain.block.repository
+package org.appjam.smashing.domain.user.repository
 
 interface BlockRepositoryCustom {
     fun findAllRelatedBlockIds(userId: String): List<String>
